add new seed script, remove old ones
