+++ seed/seed.js
@@ -1,81 +1,97 @@
 const { Types } = require('mongoose');
 const connection = require('../config/connection');
 const { User, Thought, Reaction } = require('../models');
-const {
-    getRandomNames,
-    getRandomEmail,
-    getRandomThoughts,
-    getObjectIds
-} = require('./data');
 
 connection.on('error', (err) => console.error(err));
 
 connection.once('open', async () => {
     console.log("Connected to database");
-
+    
     // Delete existing Users
     await User.deleteMany({});
 
     // Delete existing Thoughts
     await Thought.deleteMany({});
 
-    // An array to hold the Users we will be seeding
-    const users = [];
-    //const thoughts = [];
+    const thoughts = [
+        {
+            _id: new Types.ObjectId(),
+            thoughtText: "User1's first thought!",
+            username: 'user1',
+            reactions: [
+                {
+                    reactionText: "User2's reaction",
+                    username: 'user2'
+                },
+                {
+                    reactionText: "User3's reaction",
+                    username: 'user3'
+                }
+            ]
+        },
+        {
+            _id: new Types.ObjectId(),
+            thoughtText: "User2's first thought!",
+            username: 'user2',
+            reactions: [
+                {
+                    reactionText: "User1's reaction",
+                    username: 'user1'
+                },
+                {
+                    reactionText: "User3's reaction",
+                    username: 'user3'
+                }
+            ]
+        },
+        {
+            _id: new Types.ObjectId(),
+            thoughtText: "User3's first thought!",
+            username: 'user3',
+            reactions: [
+                {
+                    reactionText: "User1's reaction",
+                    username: 'user1'
+                },
+                {
+                    reactionText: "User2's reaction",
+                    username: 'user2'
+                }
+            ]
+        }
+    ];
 
-    // Create 10 example Users, each with 3 Thoughts and 2 friends
-    // HOW THIS WORKS:
-    // We cannot 
-    for (let u = 0; u < 10; u++) {
-        // create User
-        let newUser = {
-            username: getRandomNames(1),
-            email: getRandomEmail(),
-            thoughts: getObjectIds(3),
-            friends: getObjectIds(2)
-        };
+    await Thought.collection.insertMany(thoughts);
 
-        // Build our users array
-        users.push(newUser);
+    const userIds = [
+        new Types.ObjectId(),
+        new Types.ObjectId(),
+        new Types.ObjectId()
+    ]
 
-        // Create 3 Thoughts, populate with 2 random Reactions
-        
-
-        // // Seed 5 friends
-        // for (let f = 0; f < 5; f++) {
-        //     let newFriend = {
-        //         username: getRandomName(),
-        //         email: getRandomEmail(),
-        //         thoughts: [],
-        //         friends: []
-        //     };
-        //     // Place new friend into User
-        //     newUser.friends.push(newFriend);
-        // }
-
-    }
-
-    // 
-
-    const thoughts = {
-        _id: getObjectIds(1),
-        thoughtText: "if anyone finds my keys can you pls tell me? thanks",
-        username: getRandomNames(1),
-        reactions: [
-            {
-                reactionText: "wow",
-                username: getRandomNames(1)
-            },
-            {
-                reactionText: "this is not real",
-                username: getRandomNames(1)
-            },
-            {
-                reactionText: "this is a reaction",
-                username: getRandomNames(1)
-            }
-        ]
-    };
+    const users = [
+        {
+            _id: userIds[0],
+            username: "user1",
+            email: 'user1@gmail.com',
+            thoughts: [thoughts[0]._id],
+            friends: [userIds[1], userIds[2]]
+        },
+        {
+            _id: userIds[1],
+            username: "user2",
+            email: 'user2@gmail.com',
+            thoughts: [thoughts[1]._id],
+            friends: [userIds[0], userIds[2]]
+        },
+        {
+            _id: userIds[2],
+            username: "user3",
+            email: 'user3@gmail.com',
+            thoughts: [thoughts[2]._id],
+            friends: [userIds[0], userIds[1]]
+        },
+    ];
 
     await User.collection.insertMany(users);
